requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
cryptography==2.6.1
dnspython==1.16.0
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.3.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.1
marshmallow==2.19.0
openapi-spec-validator==0.2.6
pathlib==1.0.1
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
swagger-ui-bundle==0.0.3
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
